// backend/src/types/generated/public/UserSubscriptions.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { UsersId } from './Users';
import type { SubscriptionPlansId } from './SubscriptionPlans';

/** Identifier type for public.user_subscriptions */
export type UserSubscriptionsId = string & { __brand: 'public.user_subscriptions' };

/** Represents the table public.user_subscriptions */
export default interface UserSubscriptions {
  id: UserSubscriptionsId;

  user_id: UsersId | null;

  phone: string | null;

  customer_name: string | null;

  plan_id: SubscriptionPlansId | null;

  custom_items: Record<string, unknown> | null;

  monthly_price: string;

  status: string | null;

  cloudpayments_subscription_id: string | null;

  cloudpayments_token: string | null;

  current_period_start: string | null;

  current_period_end: string | null;

  next_payment_date: string | null;

  pause_until: string | null;

  cancel_reason: string | null;

  created_at: string | null;

  updated_at: string | null;
}

/** Represents the initializer for the table public.user_subscriptions */
export interface UserSubscriptionsInitializer {
  /** Default value: uuid_generate_v4() */
  id?: UserSubscriptionsId;

  user_id?: UsersId | null;

  phone?: string | null;

  customer_name?: string | null;

  plan_id?: SubscriptionPlansId | null;

  /** Default value: '[]'::jsonb */
  custom_items?: Record<string, unknown> | null;

  monthly_price: string;

  /** Default value: 'active'::character varying */
  status?: string | null;

  cloudpayments_subscription_id?: string | null;

  cloudpayments_token?: string | null;

  current_period_start?: string | null;

  current_period_end?: string | null;

  next_payment_date?: string | null;

  pause_until?: string | null;

  cancel_reason?: string | null;

  /** Default value: now() */
  created_at?: string | null;

  /** Default value: now() */
  updated_at?: string | null;
}

/** Represents the mutator for the table public.user_subscriptions */
export interface UserSubscriptionsMutator {
  id?: UserSubscriptionsId;

  user_id?: UsersId | null;

  phone?: string | null;

  customer_name?: string | null;

  plan_id?: SubscriptionPlansId | null;

  custom_items?: Record<string, unknown> | null;

  monthly_price?: string;

  status?: string | null;

  cloudpayments_subscription_id?: string | null;

  cloudpayments_token?: string | null;

  current_period_start?: string | null;

  current_period_end?: string | null;

  next_payment_date?: string | null;

  pause_until?: string | null;

  cancel_reason?: string | null;

  created_at?: string | null;

  updated_at?: string | null;
}